requirements.txt
# Core dependencies for agentic_patterns_new
python-dotenv>=1.0.0
pydantic>=2.0.0
langdetect>=1.0.9
rich>=13.0.0

# The openai package is optional.  Install it and set USE_OPENAI=true in your
# .env file if you want to call OpenAI models instead of using the built‑in
# stub.  Do not enable this unless you provide your own API key.
openai>=1.12.0